added auto aim toward turrets
The player aims up/down to shoot the enemy automatically *as long* as the Aim decal is close to an enemy (using sphere collision).   Only working on turrets right now because the way it was set up.    This doesnt work with pistols as they didnt have a "look around" animation.